$this->is_admin

+++ public/api/1.0/blocks_api.php
@@ -54,20 +54,22 @@ $app->get('/blocks/{block_id}', function (Request $request, Response $response) 
 $app->post('/blocks/', function(Request $request, Response $response) {
   $status = array();
   $status['status'] = false;
-  $post_data = $request->getParsedBody();
-  if(isset($post_data['block_name'])) {
-    //var_dump($all_gl);
-    $mysqli = $this->db;
+  if($this->is_admin) {
+    $post_data = $request->getParsedBody();
+    if(isset($post_data['block_name'])) {
+      //var_dump($all_gl);
+      $mysqli = $this->db;
 
-    if($stmt = $mysqli->prepare("INSERT INTO `blocks` (`block_id`, `block_name`) VALUES (NULL, ?);")) {
-      $stmt->bind_param("s", $post_data['block_name']);
-      $stmt->execute();
-      if($stmt->affected_rows == 1) {
-        $status['status'] = true;
+      if($stmt = $mysqli->prepare("INSERT INTO `blocks` (`block_id`, `block_name`) VALUES (NULL, ?);")) {
+        $stmt->bind_param("s", $post_data['block_name']);
+        $stmt->execute();
+        if($stmt->affected_rows == 1) {
+          $status['status'] = true;
+        }
+        $stmt->close();
+      } else {
+        echo $mysqli->error;
       }
-      $stmt->close();
-    } else {
-      echo $mysqli->error;
     }
   }
   $response->getBody()->write(json_encode($status, JSON_PRETTY_PRINT));
@@ -77,22 +79,25 @@ $app->post('/blocks/', function(Request $request, Response $response) {
 $app->put('/blocks/', function(Request $request, Response $response) {
   $status = array();
   $status['status'] = false;
-  $post_data = $request->getParsedBody();
-  //try to save data
-  if(isset($post_data['block_id']) && isset($post_data['block_name'])) {
-    $mysqli = $this->db;
+  if($this->is_admin) {
+    $post_data = $request->getParsedBody();
+    //try to save data
+    if(isset($post_data['block_id']) && isset($post_data['block_name'])) {
+      $mysqli = $this->db;
 
-    if($stmt = $mysqli->prepare("UPDATE `blocks` SET `block_name` = ? WHERE `blocks`.`block_id` = ? LIMIT 1;")) {
-      $stmt->bind_param("si", $post_data['block_name'], $post_data['block_id']);
-      $stmt->execute();
-      if($stmt->affected_rows == 1) {
-        $status['status'] = true;
+      if($stmt = $mysqli->prepare("UPDATE `blocks` SET `block_name` = ? WHERE `blocks`.`block_id` = ? LIMIT 1;")) {
+        $stmt->bind_param("si", $post_data['block_name'], $post_data['block_id']);
+        $stmt->execute();
+        if($stmt->affected_rows == 1) {
+          $status['status'] = true;
+        }
+        $stmt->close();
+      } else {
+        echo $mysqli->error;
       }
-      $stmt->close();
-    } else {
-      echo $mysqli->error;
     }
   }
+  
   $response->getBody()->write(json_encode($status, JSON_PRETTY_PRINT));
   return $response;
 });
@@ -100,20 +105,23 @@ $app->put('/blocks/', function(Request $request, Response $response) {
 $app->delete('/blocks/', function (Request $request, Response $response) {
   $resp = array();
   $resp['status'] = false;
-  $post_data = $request->getParsedBody();
+  if($this->is_admin) {
+    $post_data = $request->getParsedBody();
     $mysqli = $this->db;
-  foreach ($post_data as $key => $value) {
-    if($stmt = $mysqli->prepare("DELETE FROM `blocks` WHERE `block_id` = ?;")) {
-      $stmt->bind_param("i", $value['block_id']);
-      $stmt->execute();
-      if($stmt->affected_rows > 0) {
-        $resp['status'] = true;
+    foreach ($post_data as $key => $value) {
+      if($stmt = $mysqli->prepare("DELETE FROM `blocks` WHERE `block_id` = ?;")) {
+        $stmt->bind_param("i", $value['block_id']);
+        $stmt->execute();
+        if($stmt->affected_rows > 0) {
+          $resp['status'] = true;
+        }
+        $stmt->close();
+      } else {
+        echo $mysqli->error;
       }
-      $stmt->close();
-    } else {
-      echo $mysqli->error;
     }
   }
+  
   $response->getBody()->write(json_encode($resp, JSON_PRETTY_PRINT));
   return $response;
 });
