+++ public/api/1.0/dates_api.php
@@ -32,22 +32,25 @@ $app->get('/dates/', function (Request $request, Response $response) {
 $app->post('/dates/', function(Request $request, Response $response) {
   $status = array();
   $status['status'] = false;
-  $post_data = $request->getParsedBody();
-  if(isset($post_data['date'])) {
-    //var_dump($all_gl);
-    $mysqli = $this->db;
+  if($this->is_admin) {
+    $post_data = $request->getParsedBody();
+    if(isset($post_data['date'])) {
+      //var_dump($all_gl);
+      $mysqli = $this->db;
 
-    if($stmt = $mysqli->prepare("INSERT INTO `dates` (`date`) VALUES (?);")) {
-      $stmt->bind_param("i", $post_data['date']);
-      $stmt->execute();
-      if($stmt->affected_rows == 1) {
-        $status['status'] = true;
+      if($stmt = $mysqli->prepare("INSERT INTO `dates` (`date`) VALUES (?);")) {
+        $stmt->bind_param("i", $post_data['date']);
+        $stmt->execute();
+        if($stmt->affected_rows == 1) {
+          $status['status'] = true;
+        }
+        $stmt->close();
+      } else {
+        echo $mysqli->error;
       }
-      $stmt->close();
-    } else {
-      echo $mysqli->error;
     }
   }
+  
   $response->getBody()->write(json_encode($status, JSON_PRETTY_PRINT));
   return $response;
 });
@@ -55,22 +58,25 @@ $app->post('/dates/', function(Request $request, Response $response) {
 $app->put('/dates/', function(Request $request, Response $response) {
   $status = array();
   $status['status'] = false;
-  $post_data = $request->getParsedBody();
-  //try to save data
-  if(isset($post_data['date']) && isset($post_data['date_old'])) {
-    $mysqli = $this->db;
+  if($this->is_admin) {
+    $post_data = $request->getParsedBody();
+    //try to save data
+    if(isset($post_data['date']) && isset($post_data['date_old'])) {
+      $mysqli = $this->db;
 
-    if($stmt = $mysqli->prepare("UPDATE `dates` SET `date` = ? WHERE `dates`.`date` = ? LIMIT 1;")) {
-      $stmt->bind_param("ii", $post_data['date'], $post_data['date_old']);
-      $stmt->execute();
-      if($stmt->affected_rows == 1) {
-        $status['status'] = true;
+      if($stmt = $mysqli->prepare("UPDATE `dates` SET `date` = ? WHERE `dates`.`date` = ? LIMIT 1;")) {
+        $stmt->bind_param("ii", $post_data['date'], $post_data['date_old']);
+        $stmt->execute();
+        if($stmt->affected_rows == 1) {
+          $status['status'] = true;
+        }
+        $stmt->close();
+      } else {
+        echo $mysqli->error;
       }
-      $stmt->close();
-    } else {
-      echo $mysqli->error;
     }
   }
+  
   $response->getBody()->write(json_encode($status, JSON_PRETTY_PRINT));
   return $response;
 });
@@ -78,20 +84,23 @@ $app->put('/dates/', function(Request $request, Response $response) {
 $app->delete('/dates/', function (Request $request, Response $response) {
   $resp = array();
   $resp['status'] = false;
-  $post_data = $request->getParsedBody();
+  if($this->is_admin) {
+    $post_data = $request->getParsedBody();
     $mysqli = $this->db;
-  foreach ($post_data as $key => $value) {
-    if($stmt = $mysqli->prepare("DELETE FROM `dates` WHERE `date` = ?;")) {
-      $stmt->bind_param("i", $value['date']);
-      $stmt->execute();
-      if($stmt->affected_rows > 0) {
-        $resp['status'] = true;
+    foreach ($post_data as $key => $value) {
+      if($stmt = $mysqli->prepare("DELETE FROM `dates` WHERE `date` = ?;")) {
+        $stmt->bind_param("i", $value['date']);
+        $stmt->execute();
+        if($stmt->affected_rows > 0) {
+          $resp['status'] = true;
+        }
+        $stmt->close();
+      } else {
+        echo $mysqli->error;
       }
-      $stmt->close();
-    } else {
-      echo $mysqli->error;
     }
   }
+  
   $response->getBody()->write(json_encode($resp, JSON_PRETTY_PRINT));
   return $response;
 });
