+++ public/api/1.0/locations_api.php
@@ -57,39 +57,14 @@ $app->get('/locations/{location_id}', function (Request $request, Response $resp
 $app->post('/locations/', function(Request $request, Response $response) {
   $status = array();
   $status['status'] = false;
-  $post_data = $request->getParsedBody();
-  if(isset($post_data['location_name'])) {
-    //var_dump($all_gl);
-    $mysqli = $this->db;
+  if($this->is_admin) {
+    $post_data = $request->getParsedBody();
+    if(isset($post_data['location_name'])) {
+      //var_dump($all_gl);
+      $mysqli = $this->db;
 
-    if($stmt = $mysqli->prepare("INSERT INTO `locations` (`location_id`, `location_name`) VALUES (NULL, ?);")) {
-      $stmt->bind_param("s", $post_data['location_name']);
-      $stmt->execute();
-      if($stmt->affected_rows == 1) {
-        $status['status'] = true;
-      }
-      $stmt->close();
-    } else {
-      echo $mysqli->error;
-    }
-  }
-  $response->getBody()->write(json_encode($status, JSON_PRETTY_PRINT));
-  return $response;
-});
-
-$app->put('/locations/', function(Request $request, Response $response) {
-  $status = array();
-  $status['status'] = false;
-  $post_data = $request->getParsedBody();
-  //try to save data
-  if(isset($post_data['location_id'])) {
-    $mysqli = $this->db;
-    $old_data = GetLocation($mysqli, intval($post_data['location_id']));
-    if($old_data->location_id == intval($post_data['location_id'])) {
-      $location_name = isset($post_data['location_name']) ? $post_data['location_name'] : $old_data->location_name;
-
-      if($stmt = $mysqli->prepare("UPDATE `locations` SET `location_name` = ? WHERE `locations`.`location_id` = ? LIMIT 1;")) {
-        $stmt->bind_param("si", $location_name, $post_data['location_id']);
+      if($stmt = $mysqli->prepare("INSERT INTO `locations` (`location_id`, `location_name`) VALUES (NULL, ?);")) {
+        $stmt->bind_param("s", $post_data['location_name']);
         $stmt->execute();
         if($stmt->affected_rows == 1) {
           $status['status'] = true;
@@ -100,6 +75,37 @@ $app->put('/locations/', function(Request $request, Response $response) {
       }
     }
   }
+  
+  $response->getBody()->write(json_encode($status, JSON_PRETTY_PRINT));
+  return $response;
+});
+
+$app->put('/locations/', function(Request $request, Response $response) {
+  $status = array();
+  $status['status'] = false;
+  if($this->is_admin) {
+    $post_data = $request->getParsedBody();
+    //try to save data
+    if(isset($post_data['location_id'])) {
+      $mysqli = $this->db;
+      $old_data = GetLocation($mysqli, intval($post_data['location_id']));
+      if($old_data->location_id == intval($post_data['location_id'])) {
+        $location_name = isset($post_data['location_name']) ? $post_data['location_name'] : $old_data->location_name;
+
+        if($stmt = $mysqli->prepare("UPDATE `locations` SET `location_name` = ? WHERE `locations`.`location_id` = ? LIMIT 1;")) {
+          $stmt->bind_param("si", $location_name, $post_data['location_id']);
+          $stmt->execute();
+          if($stmt->affected_rows == 1) {
+            $status['status'] = true;
+          }
+          $stmt->close();
+        } else {
+          echo $mysqli->error;
+        }
+      }
+    }
+  }
+  
   $response->getBody()->write(json_encode($status, JSON_PRETTY_PRINT));
   return $response;
 });
@@ -107,20 +113,23 @@ $app->put('/locations/', function(Request $request, Response $response) {
 $app->delete('/locations/', function (Request $request, Response $response) {
   $resp = array();
   $resp['status'] = false;
-  $post_data = $request->getParsedBody();
+  if($this->is_admin) {
+    $post_data = $request->getParsedBody();
     $mysqli = $this->db;
-  foreach ($post_data as $key => $value) {
-    if($stmt = $mysqli->prepare("DELETE FROM `locations` WHERE `location_id` = ?;")) {
-      $stmt->bind_param("i", $value['location_id']);
-      $stmt->execute();
-      if($stmt->affected_rows > 0) {
-        $resp['status'] = true;
+    foreach ($post_data as $key => $value) {
+      if($stmt = $mysqli->prepare("DELETE FROM `locations` WHERE `location_id` = ?;")) {
+        $stmt->bind_param("i", $value['location_id']);
+        $stmt->execute();
+        if($stmt->affected_rows > 0) {
+          $resp['status'] = true;
+        }
+        $stmt->close();
+      } else {
+        echo $mysqli->error;
       }
-      $stmt->close();
-    } else {
-      echo $mysqli->error;
     }
   }
+  
   $response->getBody()->write(json_encode($resp, JSON_PRETTY_PRINT));
   return $response;
 });
